backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
gunicorn